HW_Search

+++ app/src/main/java/com/example/homework_1/ui/profile/ProfileViewModel.kt
@@ -1,7 +1,9 @@
 package com.example.homework_1.ui.profile
 
+import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.homework_1.model.User
 import com.example.homework_1.repositories.NoteRepository
 import com.example.homework_1.repositories.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -15,20 +17,24 @@ class ProfileViewModel @Inject constructor(
     private val userRepository: UserRepository
 ) : ViewModel() {
 
-    suspend fun getUserNameByEmail(email: String): String {
-        val isUserExist = userRepository.getUser(email)
-        return "${isUserExist.firstName} ${isUserExist.secondName}"
+    val user = MutableLiveData<String>()
+    val numberOfNotes = MutableLiveData<Int>()
+
+    fun getUserNameByEmail(email: String) {
+        viewModelScope.launch(Dispatchers.IO) {
+            user.postValue("${userRepository.getUser(email)?.firstName} ${userRepository.getUser(email)?.secondName}")
+        }
     }
 
-    suspend fun getUserNotesNumber(email: String): String {
-        val userNotes = noteRepository.getUserNotesByEmail(email)
-        return "${userNotes.size} notes"
+    fun getUserNotesNumber(email: String) {
+        viewModelScope.launch(Dispatchers.IO) {
+            numberOfNotes.postValue(noteRepository.getUserNotesByEmail(email).size)
+        }
     }
 
     fun deleteAllNotes(email: String) {
         viewModelScope.launch(Dispatchers.IO) {
             val userNotes = noteRepository.getUserNotesByEmail(email)
-
             for (note in userNotes) {
                 noteRepository.removeNote(note)
             }
@@ -38,7 +44,9 @@ class ProfileViewModel @Inject constructor(
     fun deleteUser(email: String) {
         viewModelScope.launch(Dispatchers.IO) {
             val user = userRepository.getUser(email)
-            userRepository.removeUser(user)
+            if (user != null) {
+                userRepository.removeUser(user)
+            }
         }
     }
 }
